Use u32 for pagination limit
(cherry picked from commit 0a9b745cd4132772f1ad682c0af0319ca7d12237)

+++ packages/std/src/pagination.rs
@@ -1,12 +1,12 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use crate::{Binary, Uint64};
+use crate::Binary;
 
 /// Simplified version of the PageRequest type for pagination from the cosmos-sdk
 #[derive(Serialize, Deserialize, Clone, Default, Debug, PartialEq, Eq, JsonSchema)]
 pub struct PageRequest {
     pub key: Option<Binary>,
-    pub limit: Uint64,
+    pub limit: u32,
     pub reverse: bool,
 }
